[cor] Rename VecInt.h to Vectors.h
