+++ code/Widgets/Widget.h
@@ -14,7 +14,7 @@
 
 #include "Core/Callbacks/CallbackList.h"
 #include "Core/Callbacks/CallbackMacro.h"
-#include "Core/Math/VecInt.h"
+#include "Core/Math/Vectors.h"
 
 #include "Widgets/Tools/Color.h"
